\name{translator}
\alias{translator}
\title{Constructor for the Translator instance}
\usage{
  translator(client_id, client_secret, default_toLang = "")
}
\arguments{
  \item{client_id}{Client ID}

  \item{client_secret}{Client secret}

  \item{default_toLang}{Default Language to output}
}
\description{
  This function generates a new Translator instance to use
  the Microsoft Translator API
}

